requirements.txt
torch>=2.0.0
tqdm>=4.64.0
matplotlib>=3.5.0
scikit-learn>=1.1.0
PyYAML>=6.0
rich>=13.0.0
numpy>=1.21.0 